Update tentative plans to set Tiny YOLO v2 as goal
tl;dr: VGG16 preprocessing is PITA, will need to reimplement

Original plan was VGG16 because it does not have many layers and is very
well-known. Unfortunately, there are issues because preprocessing for VGG-16
is supposed to be through image mean subtraction according to the paper, but
Darknet appears to ignore this (https://pjreddie.com/darknet/imagenet/). I
tried to reimplement the vgg-16 onnx model with Keras and onnxmltools, but it
will not work because vgg-16 Keras weights are trained using mean subtraction
preprocessing and therefore need preprocessed inference images. In other
words, I will need to retrain Keras and Darknet VGG16 models from scratch
to have the same image preprocessing. The layers ported over nicely but the
weights will currently not.

Tiny YOLO v2 actually seems like a better idea because YOLO is arguably the
most important and unique feature of Darknet. VGG16/19 also have more than 100
million parameters which takes a few seconds to load, so a tiny model is
better for development.